Calculate number tests
The number of tests is the product of #datasets and #of configured
test values in nodediag config file.

Before pools imported, #tests is 0.
After imported, it depends on whether the node is hosting just
one Lustre target, or hosting both (has failover target as well).

So we calculate number of tests.

0 tests is not a failure becase that is normal when node first
boots and the resource manager has not yet imported the pools.

We will need to run nodediag after pool import.